Refactor ordered_map into separate files
Separate json and xml, from the regular yaml.

Makes it possible to compile, without those...
